refactor: change logger structure
