+++ src/app.ts
@@ -6,12 +6,9 @@ import YAML from 'yamljs';
 import userRouter from './resources/users/user.router';
 import boardRouter from './resources/boards/board.router';
 import errorHandler from './middleware/errorHandler';
-import {
-  logger,
-  LogTypes,
-  parseErrorToLog,
-  serverLoggerMiddleware,
-} from './common/logger';
+import { LogTypes } from './types/logger.types';
+import { loggerMiddleware } from './middleware/logger.middleware';
+import { logger, parseErrorToLog } from './common/logger';
 
 const app = express();
 const swaggerDocument = YAML.load(
@@ -19,7 +16,7 @@ const swaggerDocument = YAML.load(
 ) as Object;
 
 app.use(express.json());
-app.use(serverLoggerMiddleware);
+app.use(loggerMiddleware);
 
 app.use('/doc', swaggerUI.serve, swaggerUI.setup(swaggerDocument));
