+++ src/types/logger.types.ts
@@ -1,0 +1,6 @@
+export enum LogTypes {
+  httpError = '[HTTP_ERROR_LOG]',
+  httpSuccess = '[HTTP_SUCCESS_LOG]',
+  uncaughtExceptionError = '[UNCAUGHT_EXCEPTION_ERROR]',
+  unhandledRejectionError = '[UNCAUGHT_REJECTION_ERROR]',
+}
